Create new folder and split classes into it
