+++ extensions/ql-vscode/test/vscode-tests/cli-integration/queries.test.ts
@@ -19,7 +19,7 @@ import { CliVersionConstraint, CodeQLCliServer } from "../../../src/cli";
 import { describeWithCodeQL } from "../cli";
 import { CoreCompletedQuery, QueryRunner } from "../../../src/queryRunner";
 import { SELECT_QUERY_NAME } from "../../../src/contextual/locationFinder";
-import { LocalQueries } from "../../../src/local-queries";
+import { LocalQueries } from "../../../src/local-queries/local-queries";
 import { QueryResultType } from "../../../src/pure/new-messages";
 import { createVSCodeCommandManager } from "../../../src/common/vscode/commands";
 import {
